requirements.txt
requests~=2.29.0
pandas~=2.0.1
python-dateutil~=2.8.2
pytz~=2023.3
shapely~=2.0.1
osmnx~=1.3.0